services/reddit-scraper/requirements.txt
psycopg2-binary
praw
python-dotenv